# api/openapi.yaml
openapi: 3.0.1
info:
  title: ToDo Manager - 25.02 Exam Call
  description: REST APIs for the interaction with the ToDo Manager service, extended with Collaborative tasks.
  contact:
    name: Lino Mediavilla (s293655)
    url: https://github.com/linomp
  version: 2.0.0
servers:
  - url: "http://localhost:3000"
    description: Local Dev server
...
paths:
  /api/tasks/public:
    get:
      tags:
        - tasks/public
      summary: Retrieve the public tasks
      description: >-
        The public tasks (i.e., the tasks that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicTasks
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: "The id of the requested page (if absent, the first page is returned)"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  next:
                    type: string
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
  /api/tasks:
    post:
      tags:
        - tasks
      security:
        - cookieAuth: []
      summary: Create a new task
      description: A new task is created by the authenticated user.
      operationId: addTask
      requestBody:
        description: Task object that needs to be created. In particular, now a **minimum number of completers** can be specified for a Task (default value is 1). This indicates **the number of assignees that must mark the task as completed** before the task itself can be considered complete.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
            examples:
              collaborative task:
                summary: example collaborative task
                $ref: "#/components/examples/collab_task_1"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "500":
          description: Internal Server Error
          content: {}
...
  "/api/tasks/{taskId}":
    put:
      tags:
        - "tasks/{taskId}"
      security:
        - cookieAuth: []
      summary: Update a task
      description: The task with ID taskId is updated. This operation does not allow to mark a task as completed. This operation can be performed only by the owner.
      operationId: updateSingleTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The updated task object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
        required: true
      responses:
        "204":
          description: No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
    get:
      tags:
        - "tasks/{taskId}"
      security:
        - cookieAuth: []
      summary: Retrieve a task
      description: The task with ID taskId is retrieved. This operation can be performed on the task if at least one of the following three conditions is satisfied. 1) The task is public. 2) The user who performs the operation is the task's owner. 3) The user who performs the operation is a task's assignee.
      operationId: getSingleTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
              examples:
                completed task:
                  summary: "completed task"
                  $ref: "#/components/examples/completed_task"
                partially completed task:
                  summary: "partially completed task"
                  $ref: "#/components/examples/non_completed_task"
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
    delete:
      tags:
        - "tasks/{taskId}"
      security:
        - cookieAuth: []
      summary: Delete a task
      description: The task with ID taskId is deleted. This operation can only be performed by the owner.
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
  /api/users:
    get:
      tags:
        - users
      security:
        - cookieAuth: []
      summary: Get information about the users
      description: >-
        The available information (passwords excluded) about all the
        users is retrieved.
      operationId: getUsers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
...
  "/api/tasks/{taskId}/completion":
    put:
      tags:
        - "tasks/{taskId}/completion"
      security:
        - cookieAuth: []
      summary: Complete a task
      description: >-
        **Increases** the number of **task assignees that have completed the task** with ID taskId.
        If the **minimum number of completers** is **reached**, the task is also **marked as completed**. 
        This operation can **only be performed by an assignee of the task**, and only if they **have not marked the task as completed before**.
      operationId: completeTask
      parameters:
        - name: taskId
          in: path
          description: ID of the task
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content:
            application/json:
              examples:
                repeated complete request:
                  value:
                    {
                      "errors":
                        [
                          {
                            "param": "Server",
                            "msg": "A user cannot mark the activity as completed more than once",
                          },
                        ],
                    }
                already completed activity:
                  value:
                    {
                      "errors":
                        [
                          {
                            "param": "Server",
                            "msg": "Activity is already completed",
                          },
                        ],
                    }
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
  "/api/users/{userId}/tasks/created":
    get:
      tags:
        - users/{userId}/tasks/created
      security:
        - cookieAuth: []
      summary: Retrieve the tasks created by the user with ID userId
      description: >-
        This operation allows an authenticated user to retrieve the tasks that she has created. A pagination mechanism is implemented to limit the size of messages.
      operationId: getUserTasks
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  next:
                    type: string
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
  "/api/users/{userId}/tasks/assigned":
    get:
      tags:
        - users/{userId}/tasks/assigned
      security:
        - cookieAuth: []
      summary: Retrieve the tasks assigned to the user with ID userId
      description: >-
        This operation allows an authenticated user to retrieve the tasks that has been assigned to her. A pagination mechanism is implemented to limit the size of messages.
      operationId: getAssignedTasks
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: The id of the requested page (if absent, the first page is returned)
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  tasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
                  next:
                    type: string
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
...
components:
  schemas:
    User:
      type: object
      properties:
        $schema:
          type: string
          description: >-
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the user
        name:
          type: string
          description: username of the user
        email:
          type: string
          format: email
          description: email address of the user
        password:
          type: string
          minLength: 6
          maxLength: 20
          description: password of the user
        self:
          type: string
          format: uri-reference
          description: self link
    Task:
      type: object
      properties:
        $schema:
          type: string
          description: >-
            path to the JSON Schema to use for the validation of the JSON file
            describing the task
        id:
          type: integer
          description: unique identifier for the task
        description:
          type: string
          maxLength: 160
          description: textual task description
        important:
          type: boolean
          description: >-
            this property is set to true if the task is marked as important,
            false otherwise
          default: false
        private:
          type: boolean
          description: >-
            this property is set to true if the task is marked as private, false
            otherwise
          default: true
        project:
          type: string
          description: project in which the task is inserted
        deadline:
          type: string
          format: date-time
          description: the due date and hour of the task
        completed:
          type: boolean
          description: >-
            this property is set to true if the task is marked as completed,
            false otherwise
          default: false
        owner:
          $ref: "#/components/schemas/User"
        assignees:
          type: array
          items:
            $ref: "#/components/schemas/User"
        completers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        selectors:
          type: array
          items:
            $ref: "#/components/schemas/User"
        minCompleters:
          type: integer
          default: 1
        currentCompleters:
          type: integer
          default: 0
        self:
          type: string
          format: uri-reference
          description: self link
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid
  examples:
...
    collab_task_1:
      value:
        {
          "description": "Collaborative Task",
          "important": false,
          "private": false,
          "project": "Personal",
          "deadline": "2022-06-11T02:06:05.629Z",
          "completed": false,
          "minCompleters": 2,
        }
    completed_task:
      value:
        {
          "id": 25,
          "description": "Collaborative Task",
          "important": false,
          "private": false,
          "deadline": "2022-06-11T02:06:05.629Z",
          "project": "Personal",
          "minCompleters": 2,
          "currentCompleters": 2,
          "completed": true,
          "self": "/api/tasks/25",
        }
    non_completed_task:
      value:
        {
          "id": 27,
          "description": "Collaborative Task",
          "important": false,
          "private": false,
          "deadline": "2022-06-11T02:06:05.629Z",
          "project": "Personal",
          "minCompleters": 2,
          "currentCompleters": 1,
          "completed": false,
          "self": "/api/tasks/27",
        }
...